Add files via upload
A network packet analyzer, often referred to as a packet sniffer or network analyzer, is a tool used to capture, analyze, and interpret data packets traveling across a network. It plays a crucial role in network management, security, and troubleshooting by providing detailed insights into the data being transmitted and received over a network. Understanding network packet analyzers is important for anyone involved in network administration, cybersecurity, or IT support.